backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
redis==5.0.1
celery==5.3.4
openai==1.3.0
requests==2.31.0
django-filter==23.3
drf-yasg==1.21.7
gunicorn==21.2.0
python-dotenv==1.0.0